style(changed ui and some other fixed): style: change UI and add Arabic translation
transformed some steps into one another
