+++ localization/ar/ar.json
@@ -18,7 +18,7 @@
   "Field of Study": "مجال الدراسة",
   "Academic Level": "المستوى الأكاديمي",
   "Organization": "المنظمة",
-  "Experience and interest": "الخبرة والاهتمامات",
+  "Law form and interest": "شكل القانون والفائدة",
   "Relevant Laws and Judgments in the Application": "القوانين والأحكام ذات الصلة في التطبيق",
   "Duration": "المدة",
   "Are you currently enrolled in a legal training program?": "هل أنت مسجل حاليًا في برنامج تدريب قانوني؟",
