+++ localization/en/en.json
@@ -8,7 +8,7 @@
   "5": "5",
   "Full Name": "Full Name",
   "select language": "select language",
-  "If yes, Please specify the goal orginzatio nwhere you are training": "If yes, Please specify the goal orginzatio nwhere you are training",
+  "If yes, Please specify the goal organization where you are training": "If yes, Please specify the goal organization where you are training",
   "arabic": "arabic",
   "If yes, Please specify the legal organization where you are training": "If yes, Please specify the legal organization where you are training",
   "Do you currently enrolled in a legal training program?": "Do you currently enrolled in a legal training program?",
@@ -18,14 +18,15 @@
   "Field of Study": "Field of Study",
   "Academic Level": "Academic Level",
   "Organization": "Organization",
-  "Experience and interest": "Experience and interest",
+  "Law form and interest": "Law form and interest",
   "Relevant Laws and Judgments in the Application": "Relevant Laws and Judgments in the Application",
   "Duration": "Duration",
+  "Are you currently enrolled in a legal training program?": "Are you currently enrolled in a legal training program?",
   "Do you have specific interests in developing particular skills such as drafting legal memos, case management, or advocacy ?": "Do you have specific interests in developing particular skills such as drafting legal memos, case management, or advocacy ?",
   "Do you have practical experience in the legal field?": "Do you have practical experience in the legal field?",
   "Type of Legal Work": "Type of Legal Work",
   "Which area of law interests you the most": "Which area of law interests you the most",
-  "If yes, Please describe your interest": "",
+  "If yes, Please describe your interest": "If yes, Please describe your interest",
   "Yes": "Yes",
   "Provide-Information": "Provide-Information",
   "Moderate": "Moderate",
@@ -39,11 +40,11 @@
   "Limited": "Limited",
   "Commitment and Collaboration": "Commitment and Collaboration",
   "Very Good": "Very Good",
-  "Your Opinion on the Ai Legal Traning System": "Your Opinion on the Ai Legal Traning System",
+  "Your Opinion on the Ai Legal Training System": "Your Opinion on the Ai Legal Training System",
   "Please provide the information required in the below form": "Please provide the information required in the below form",
   "No": "No",
-  "What is your level of familiarity with AI technologies in the legal field ?": "What is your level of familiarity with AI technologies in the legal field ?",
-  "Which features do you think would be most useful in the application ?": "Which features do you think would be most useful in the application ?",
+  "What is your level of familiarity with AI technologies in the legal field?": "What is your level of familiarity with AI technologies in the legal field?",
+  "Which features do you think would be most useful in the application?": "Which features do you think would be most useful in the application?",
   "Do you see the importance of having a system to track practical training hours?": "Do you see the importance of having a system to track practical training hours?",
   "Next": "Next",
   "Previous": "Previous",
@@ -51,6 +52,7 @@
   "Periodic progress reports": "Periodic progress reports",
   "Managing training tasks between lawyers and trainees": "Managing training tasks between lawyers and trainees",
   "Other (please specify)": "Other (please specify)",
+  "Other": "Other",
   "Entering Basic Legal Laws and Judgement: Law": "Entering Basic Legal Laws and Judgement: Law",
   "Law Name": "Law Name",
   "Article Number": "Article Number",
@@ -71,7 +73,10 @@
   "Please describe your career goals": "Please describe your career goals",
   "Breif Description of the Judgment or Law": "Breif Description of the Judgment or Law",
   "Practical Examplaes of the Judgment": "Practical Examplaes of the Judgment",
+  "Judgment summary is required": "Judgment summary is required",
+  "Key highlights are required": "Key highlights are required",
+  "Practical examples are required": "Practical examples are required",
   "How much time can you dedicate daily to using the application?": "How much time can you dedicate daily to using the application?",
-  "Are you willing to participate in the application development phases by providing feddback or testing early versions ?": "Are you willing to participate in the application development phases by providing feddback or testing early versions ?",
-  "Do you have suggestions on the best ways to encourage students and trainees to use the application?": "Do you have suggestions on the best ways to encourage students and trainees to use the application?"
+  "Are you willing to participate in the application development phases?": "Are you willing to participate in the application development phases?",
+  "Do you have suggestions for encouraging students to use the application?": "Do you have suggestions for encouraging students to use the application?"
 }
